RP: Rapport: Markera rader vars totala kostnad överstiger gränsvärde eller percentil

+++ EkoStatRp/Models/ReportSettingsViewModel.cs
@@ -6,7 +6,7 @@ namespace EkoStatRp.Models;
 public class ReportSettingsViewModel
 {
     [Display(Name = "Cost limit")]
-    public double? CostLimit { get; set; }
+    public decimal? CostLimit { get; set; }
 
     [Display(Name = "Type")]
     public LimitType CostLimitType { get; set; }
